@@ -51,7 +51,7 @@ JNIEXPORT void JNICALL
 Java_xyz_asdasd_gles3jni_GLES3JNILib_resize(JNIEnv *env, jobject obj, jint width, jint height) {
 	screenWidth = width;
 	screenHeight = height;
-	game.onResize(width, height);
+	game.resize(width, height);
 }
 
 JNIEXPORT void JNICALL
@@ -62,12 +62,12 @@ Java_xyz_asdasd_gles3jni_GLES3JNILib_step(JNIEnv *env, jobject obj) {
 
 JNIEXPORT void JNICALL
 Java_xyz_asdasd_gles3jni_GLES3JNILib_onMove(JNIEnv *env, jobject obj, jfloat dx, jfloat dy) {
-	game.onMove(dx, dy);
+	game.getState()->move(dx, dy);
 }
 
 JNIEXPORT void JNICALL
 Java_xyz_asdasd_gles3jni_GLES3JNILib_onPress(JNIEnv *env, jobject obj, jfloat x, jfloat y) {
 	x -= screenWidth/2;
 	y = screenHeight/2 - y;
-	game.onPress(x, y);
+	game.getState()->press(x, y);
 }
